Reorganised code to avoid circular imports
